FEAT : 자동 로그인 로직 수정
